Extract amdkfd utilities to its own module
